# services/tracker/openapi.json
{
  "components": {
    "schemas": {
...
      "BenchmarkStatus": {
        "enum": [
          "IN_PROGRESS",
          "STOPPING",
          "STOPPED",
          "FINISHED",
          "ERROR"
        ],
        "title": "BenchmarkStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SingleBenchmarkResponse": {
        "description": "Single-run view: BenchmarkTableRow fields plus optional final_score.",
        "properties": {
          "agent_name": {
            "title": "Agent Name",
            "type": "string"
          },
          "cloudwatch_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloudwatch Url"
          },
          "error_message": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "final_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Score"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "finished_tasks": {
            "title": "Finished Tasks",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "runtime": {
            "enum": [
              "managed",
              "legacy"
            ],
            "title": "Runtime",
            "type": "string"
          },
          "s3_bucket_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "S3 Bucket Url"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "started_by_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started By Email"
          },
          "status": {
            "$ref": "#/components/schemas/BenchmarkStatus"
          },
          "task_state_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "default": {},
            "title": "Task State Counts",
            "type": "object"
          },
          "total_tasks": {
            "title": "Total Tasks",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "agent_name",
          "model",
          "started_at",
          "finished_at",
          "status",
          "total_tasks",
          "finished_tasks",
          "runtime"
        ],
        "title": "SingleBenchmarkResponse",
        "type": "object"
      },
...
      "TaskStatus": {
        "enum": [
          "PENDING",
          "BUILDING",
          "IN_PROGRESS",
          "EVALUATING",
          "STOPPED",
          "FINISHED",
          "ERROR"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
      "TaskSummary": {
        "properties": {
          "error_message": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_id",
          "status",
          "started_at",
          "finished_at"
        ],
        "title": "TaskSummary",
        "type": "object"
      },
      "TasksResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            },
            "title": "Tasks",
            "type": "array"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "tasks",
          "total_count"
        ],
        "title": "TasksResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "DescopeAccessKey": {
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      },
      "DescopeBearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/benchmarks/{benchmark_id}": {
      "get": {
        "description": "Fetch a single benchmark with task counts + final score for the SingleRun page.",
        "operationId": "get_single_benchmark",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Benchmark Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleBenchmarkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Single Benchmark"
      }
    },
...
    "/benchmarks/{benchmark_id}/tasks": {
      "get": {
        "description": "Paginated tasks for a benchmark, with optional status filter + task-id search.\n\nsort=status desc surfaces errors first (attention priority). Default: started_at desc.",
        "operationId": "get_benchmark_tasks",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Benchmark Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "task_id_search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id Search"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "started_at",
              "enum": [
                "task_id",
                "started_at",
                "duration",
                "status"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_dir",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Sort Dir",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TasksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Benchmark Tasks"
      }
    },
...
  }
}